+++ packages/lit-html/src/test/lit-html_test.ts
@@ -25,6 +25,9 @@ import {
   stripExpressionMarkers,
 } from './test-utils/strip-markers.js';
 
+const ua = window.navigator.userAgent;
+const isIe = ua.indexOf('Trident/') > 0;
+
 suite('lit-html', () => {
   let container: HTMLDivElement;
 
@@ -75,7 +78,10 @@ suite('lit-html', () => {
 
     test('text after self-closing tag', () => {
       assertRender(html`<input />${'A'}`, '<input>A');
-      assertRender(html`<x-foo />${'A'}`, '<x-foo>A</x-foo>');
+      assertRender(
+        html`<!-- @ts-ignore --><x-foo />${'A'}`,
+        '<!-- @ts-ignore --><x-foo>A</x-foo>'
+      );
     });
 
     test('text child of element with unquoted attribute', () => {
@@ -97,6 +103,22 @@ suite('lit-html', () => {
 
     test('renders multiple parts per element, preserving whitespace', () => {
       assertRender(html`<div>${'foo'} ${'bar'}</div>`, '<div>foo bar</div>');
+    });
+
+    test('renders templates with comments', () => {
+      // prettier-ignore
+      assertRender(html`
+        <div>
+          <!-- this is a comment -->
+          <h1 class="${'foo'}">title</h1>
+          <p>${'foo'}</p>
+        </div>`, `
+        <div>
+          <!-- this is a comment -->
+          <h1 class="foo">title</h1>
+          <p>foo</p>
+        </div>`
+      );
     });
 
     test('text after element', () => {
@@ -209,6 +231,115 @@ suite('lit-html', () => {
       // prettier-ignore
       assertRender(html`<scriptx>${'foo'}</scriptx>`, '<scriptx>foo</scriptx>');
     });
+
+    test('attribute after raw text element', () => {
+      // prettier-ignore
+      assertRender(
+        html`<script></script><div a=${'A'}></div>`,
+        '<script></script><div a="A"></div>'
+      );
+    });
+
+    test('unquoted attribute', () => {
+      // prettier-ignore
+      assertRender(html`<div a=${'A'}></div>`, '<div a="A"></div>');
+      // prettier-ignore
+      assertRender(html`<div abc=${'A'}></div>`, '<div abc="A"></div>');
+      // prettier-ignore
+      assertRender(html`<div abc = ${'A'}></div>`, '<div abc="A"></div>');
+    });
+
+    test('quoted attribute', () => {
+      // prettier-ignore
+      assertRender(html`<div a="${'A'}"></div>`, '<div a="A"></div>');
+      // prettier-ignore
+      assertRender(html`<div abc="${'A'}"></div>`, '<div abc="A"></div>');
+      // prettier-ignore
+      assertRender(html`<div abc = "${'A'}"></div>`, '<div abc="A"></div>');
+    });
+
+    test('second quoted attribute', () => {
+      // prettier-ignore
+      assertRender(
+        html`<div a="b" c="${'A'}"></div>`,
+        '<div a="b" c="A"></div>'
+      );
+    });
+
+    test('two quoted attributes', () => {
+      // prettier-ignore
+      assertRender(
+        html`<div a="${'A'}" b="${'A'}"></div>`,
+        '<div a="A" b="A"></div>'
+      );
+    });
+
+    test('two unquoted attributes', () => {
+      // prettier-ignore
+      assertRender(
+        html`<div a=${'A'} b=${'A'}></div>`,
+        '<div a="A" b="A"></div>'
+      );
+    });
+
+    test('quoted attribute multi', () => {
+      assertRender(html`<div a="${'A'} ${'A'}"></div>`, '<div a="A A"></div>');
+    });
+
+    test('quoted attribute with markup', () => {
+      // prettier-ignore
+      assertRender(
+        html`<div a="<table>${'A'}"></div>`,
+        '<div a="<table>A"></div>'
+      );
+    });
+
+    test('text after quoted attribute', () => {
+      assertRender(html`<div a="${'A'}">${'A'}</div>`, '<div a="A">A</div>');
+    });
+
+    test('text after unquoted attribute', () => {
+      assertRender(html`<div a=${'A'}>${'A'}</div>`, '<div a="A">A</div>');
+    });
+
+    // test('inside start tag', () => {
+    //   assertRender(html`<div ${attr`a="b"`}></div>`, '<div a="b"></div>');
+    // });
+
+    // test('inside start tag x2', () => {
+    //   // We don't support multiple attribute-position bindings yet, so just
+    //   // ensure this parses ok
+    //   assertRender(
+    //     html`<div ${attr`a="b"`} ${attr`c="d"`}></div>`,
+    //     '<div a="b"></div>'
+    //   );
+    // });
+
+    // test('inside start tag after unquoted attribute', () => {
+    //   // prettier-ignore
+    //   assertRender(html`<div a=b ${attr`c="d"`}></div>`, '<div a="b" c="d"></div>');
+    // });
+
+    // test('inside start tag after quoted attribute', () => {
+    //   // prettier-ignore
+    //   assertRender(html`<div a="b" ${attr`c="d"`}></div>`, '<div a="b" c="d"></div>');
+    // });
+
+    // test('inside start tag before unquoted attribute', () => {
+    //   // bound attributes always appear after static attributes
+    //   assertRender(
+    //     html`<div ${attr`c="d"`} a="b"></div>`,
+    //     '<div a="b" c="d"></div>'
+    //   );
+    // });
+
+    // test('inside start tag before quoted attribute', () => {
+    //   // bound attributes always appear after static attributes
+    //   assertRender(
+    //     html`<div ${attr`c="d"`} a="b"></div>`,
+    //     '<div a="b" c="d"></div>'
+    //   );
+    // });
 
     test('"dynamic" tag name', () => {
       render(html`<${'A'}></${'A'}>`, container);
@@ -383,5 +514,628 @@ suite('lit-html', () => {
       assert.equal(line.tagName, 'line');
       assert.equal(line.namespaceURI, 'http://www.w3.org/2000/svg');
     });
+  });
+
+  suite('attributes', () => {
+    test('renders to a quoted attribute', () => {
+      render(html`<div foo="${'bar'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar"></div>'
+      );
+    });
+
+    test('renders to an unquoted attribute', () => {
+      render(html`<div foo=${'bar'}></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar"></div>'
+      );
+    });
+
+    test('renders interpolation to an attribute', () => {
+      render(html`<div foo="A${'B'}C"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="ABC"></div>'
+      );
+    });
+
+    test('renders multiple bindings in an attribute', () => {
+      render(html`<div foo="a${'b'}c${'d'}e"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="abcde"></div>'
+      );
+    });
+
+    test('renders two attributes on one element', () => {
+      const result = html`<div a="${1}" b="${2}"></div>`;
+      render(result, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div a="1" b="2"></div>'
+      );
+    });
+
+    test('renders multiple bindings in two attributes', () => {
+      render(
+        html`<div foo="a${'b'}c${'d'}e" bar="a${'b'}c${'d'}e"></div>`,
+        container
+      );
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="abcde" bar="abcde"></div>'
+      );
+    });
+
+    test.skip('renders a Symbol to an attribute', () => {
+      render(html`<div foo=${Symbol('A')}></div>`, container);
+      assert.include(
+        container.querySelector('div')!.getAttribute('foo')!.toLowerCase(),
+        'symbol'
+      );
+    });
+
+    test.skip('renders a Symbol in an array to an attribute', () => {
+      render(html`<div foo=${[Symbol('A')] as any}></div>`, container);
+      assert.include(
+        container.querySelector('div')!.getAttribute('foo')!.toLowerCase(),
+        'symbol'
+      );
+    });
+
+    test('renders a binding in a style attribute', () => {
+      const t = html`<div style="color: ${'red'}"></div>`;
+      render(t, container);
+      if (isIe) {
+        assert.equal(
+          stripExpressionComments(container.innerHTML),
+          '<div style="color: red;"></div>'
+        );
+      } else {
+        assert.equal(
+          stripExpressionComments(container.innerHTML),
+          '<div style="color: red"></div>'
+        );
+      }
+    });
+
+    test('renders multiple bindings in a style attribute', () => {
+      const t = html`<div style="${'color'}: ${'red'}"></div>`;
+      render(t, container);
+      if (isIe) {
+        assert.equal(
+          stripExpressionComments(container.innerHTML),
+          '<div style="color: red;"></div>'
+        );
+      } else {
+        assert.equal(
+          stripExpressionComments(container.innerHTML),
+          '<div style="color: red"></div>'
+        );
+      }
+    });
+
+    test('renders a binding in a class attribute', () => {
+      render(html`<div class="${'red'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div class="red"></div>'
+      );
+    });
+
+    test('renders a binding in an input value attribute', () => {
+      render(html`<input value="${'the-value'}" />`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<input value="the-value">'
+      );
+      assert.equal(container.querySelector('input')!.value, 'the-value');
+    });
+
+    test('renders a case-sensitive attribute', () => {
+      const size = 100;
+      render(html`<svg viewBox="0 0 ${size} ${size}"></svg>`, container);
+      assert.include(
+        stripExpressionComments(container.innerHTML),
+        'viewBox="0 0 100 100"'
+      );
+
+      // Make sure non-alpha valid attribute name characters are handled
+      render(html`<svg view_Box="0 0 ${size} ${size}"></svg>`, container);
+      assert.include(
+        stripExpressionComments(container.innerHTML),
+        'view_Box="0 0 100 100"'
+      );
+    });
+
+    test('renders to an attribute expression after an attribute literal', () => {
+      render(html`<div a="b" foo="${'bar'}"></div>`, container);
+      // IE and Edge can switch attribute order!
+      assert.include(
+        ['<div a="b" foo="bar"></div>', '<div foo="bar" a="b"></div>'],
+        stripExpressionComments(container.innerHTML)
+      );
+    });
+
+    test('renders to an attribute expression before an attribute literal', () => {
+      render(html`<div foo="${'bar'}" a="b"></div>`, container);
+      // IE and Edge can switch attribute order!
+      assert.include(
+        ['<div a="b" foo="bar"></div>', '<div foo="bar" a="b"></div>'],
+        stripExpressionComments(container.innerHTML)
+      );
+    });
+
+    // Regression test for exception in template parsing caused by attributes
+    // reordering when a attribute binding precedes an attribute literal.
+    test('renders attribute binding after attribute binding that moved', () => {
+      render(
+        html`<a href="${'foo'}" class="bar"><div id=${'a'}></div></a>`,
+        container
+      );
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        `<a class="bar" href="foo"><div id="a"></div></a>`
+      );
+    });
+
+    test('renders a bound attribute without quotes', () => {
+      render(html`<div foo=${'bar'}></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar"></div>'
+      );
+    });
+
+    test('renders multiple bound attributes', () => {
+      render(
+        html`<div foo="${'Foo'}" bar="${'Bar'}" baz=${'Baz'}></div>`,
+        container
+      );
+      assert.oneOf(stripExpressionComments(container.innerHTML), [
+        '<div foo="Foo" bar="Bar" baz="Baz"></div>',
+        '<div foo="Foo" baz="Baz" bar="Bar"></div>',
+        '<div bar="Bar" foo="Foo" baz="Baz"></div>',
+      ]);
+    });
+
+    test('renders multiple bound attributes without quotes', () => {
+      render(
+        html`<div foo=${'Foo'} bar=${'Bar'} baz=${'Baz'}></div>`,
+        container
+      );
+      assert.oneOf(stripExpressionComments(container.innerHTML), [
+        '<div foo="Foo" bar="Bar" baz="Baz"></div>',
+        '<div foo="Foo" baz="Baz" bar="Bar"></div>',
+        '<div bar="Bar" foo="Foo" baz="Baz"></div>',
+      ]);
+    });
+
+    test('renders multi-expression attribute without quotes', () => {
+      render(html`<div foo="${'Foo'}${'Bar'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="FooBar"></div>'
+      );
+    });
+
+    test('renders to attributes with attribute-like values', () => {
+      render(html`<div foo="bar=${'foo'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar=foo"></div>'
+      );
+    });
+
+    test('does not call a function bound to an attribute', () => {
+      const f = () => {
+        throw new Error();
+      };
+      render(html`<div foo=${f as any}></div>`, container);
+      const div = container.querySelector('div')!;
+      assert.isTrue(div.hasAttribute('foo'));
+    });
+
+    test('renders an array to an attribute', () => {
+      render(html`<div foo=${[1, 2, 3] as any}></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="1,2,3"></div>'
+      );
+    });
+
+    test('renders to an attribute before a node', () => {
+      render(html`<div foo="${'bar'}">${'baz'}</div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="bar">baz</div>'
+      );
+    });
+
+    test('renders to an attribute after a node', () => {
+      // prettier-ignore
+      render(html`<div>${'baz'}</div><div foo="${'bar'}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>baz</div><div foo="bar"></div>'
+      );
+    });
+
+    test('renders undefined in attributes', () => {
+      render(html`<div attribute="it's ${undefined}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div attribute="it\'s "></div>'
+      );
+    });
+
+    test('renders undefined in attributes', () => {
+      render(html`<div attribute="${undefined as any}"></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div attribute=""></div>'
+      );
+    });
+
+    test('nothing sentinel removes an attribute', () => {
+      const go = (v: any) => html`<div a=${v}></div>`;
+      render(go(nothing), container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+
+      render(go('a'), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div a="a"></div>'
+      );
+
+      render(go(nothing), container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+    });
+
+    test('interpolated nothing sentinel removes an attribute', () => {
+      const go = (v: any) => html`<div a="A${v}"></div>`;
+      render(go('a'), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div a="Aa"></div>'
+      );
+
+      render(go(nothing), container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+    });
+
+    test('noChange works', () => {
+      const go = (v: any) => render(html`<div foo=${v}></div>`, container);
+      go('A');
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="A"></div>',
+        'A'
+      );
+      const observer = new MutationObserver(() => {});
+      observer.observe(container, {attributes: true, subtree: true});
+
+      go(noChange);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="A"></div>',
+        'B'
+      );
+      assert.isEmpty(observer.takeRecords());
+    });
+
+    test('noChange works on one of multiple expressions', () => {
+      const go = (a: any, b: any) =>
+        render(html`<div foo="${a}:${b}"></div>`, container);
+      go('A', 'B');
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="A:B"></div>',
+        'A'
+      );
+      go(noChange, 'C');
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo="A:C"></div>',
+        'B'
+      );
+    });
+  });
+
+  suite('boolean attributes', () => {
+    test('adds attributes for true values', () => {
+      render(html`<div ?foo=${true}></div>`, container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo=""></div>'
+      );
+    });
+
+    test('removes attributes for false values', () => {
+      render(html`<div ?foo=${false}></div>`, container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+    });
+
+    test('removes attributes for nothing values', () => {
+      const go = (v: any) => render(html`<div ?foo=${v}></div>`, container);
+
+      go(nothing);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+
+      go(true);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo=""></div>'
+      );
+
+      go(nothing);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+    });
+
+    test('noChange works', () => {
+      const go = (v: any) => render(html`<div ?foo=${v}></div>`, container);
+      go(true);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo=""></div>'
+      );
+      const observer = new MutationObserver(() => {});
+      observer.observe(container, {attributes: true, subtree: true});
+      go(noChange);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div foo=""></div>'
+      );
+      assert.isEmpty(observer.takeRecords());
+    });
+  });
+
+  suite('properties', () => {
+    test('sets properties', () => {
+      render(html`<div .foo=${123} .Bar=${456}></div>`, container);
+      const div = container.querySelector('div')!;
+      assert.strictEqual((div as any).foo, 123);
+      assert.strictEqual((div as any).Bar, 456);
+    });
+
+    test('nothing becomes undefined', () => {
+      const go = (v: any) => render(html`<div .foo=${v}></div>`, container);
+
+      go(1);
+      const div = container.querySelector('div')!;
+      assert.strictEqual((div as any).foo, 1);
+
+      go(nothing);
+      assert.strictEqual((div as any).foo, undefined);
+    });
+
+    test('noChange works', () => {
+      const go = (v: any) => render(html`<div .foo=${v}></div>`, container);
+      go(1);
+      const div = container.querySelector('div')!;
+      assert.strictEqual((div as any).foo, 1);
+
+      go(noChange);
+      assert.strictEqual((div as any).foo, 1);
+    });
+  });
+
+  suite('updates', () => {
+    let container: HTMLElement;
+
+    setup(() => {
+      container = document.createElement('div');
+    });
+
+    test('dirty checks simple values', () => {
+      const foo = 'aaa';
+
+      const t = () => html`<div>${foo}</div>`;
+
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>aaa</div>'
+      );
+      const text = container.querySelector('div')!;
+      assert.equal(text.textContent, 'aaa');
+
+      // Set textContent manually. Since lit-html doesn't dirty check against
+      // actual DOM, but again previous part values, this modification should
+      // persist through the next render with the same value.
+      text.textContent = 'bbb';
+      assert.equal(text.textContent, 'bbb');
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>bbb</div>'
+      );
+
+      // Re-render with the same content, should be a no-op
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>bbb</div>'
+      );
+      const text2 = container.querySelector('div')!;
+
+      // The next node should be the same too
+      assert.strictEqual(text, text2);
+    });
+
+    test('dirty checks node values', async () => {
+      const node = document.createElement('div');
+      const t = () => html`${node}`;
+
+      const observer = new MutationObserver(() => {});
+      observer.observe(container, {childList: true, subtree: true});
+
+      assert.equal(stripExpressionComments(container.innerHTML), '');
+      render(t(), container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+
+      const elementNodes: Node[] = [];
+      let mutationRecords: MutationRecord[] = observer.takeRecords();
+      for (const record of mutationRecords) {
+        elementNodes.push(
+          ...Array.from(record.addedNodes).filter(
+            (n) => n.nodeType === Node.ELEMENT_NODE
+          )
+        );
+      }
+      assert.equal(elementNodes.length, 1);
+
+      mutationRecords = [];
+      render(t(), container);
+      assert.equal(stripExpressionComments(container.innerHTML), '<div></div>');
+      mutationRecords = observer.takeRecords();
+      assert.equal(mutationRecords.length, 0);
+    });
+
+    test('renders to and updates a container', () => {
+      let foo = 'aaa';
+
+      const t = () => html`<div>${foo}</div>`;
+
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>aaa</div>'
+      );
+      const div = container.querySelector('div')!;
+      assert.equal(div.tagName, 'DIV');
+
+      foo = 'bbb';
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>bbb</div>'
+      );
+      const div2 = container.querySelector('div')!;
+      // check that only the part changed
+      assert.equal(div, div2);
+    });
+
+    test('renders to and updates sibling parts', () => {
+      let foo = 'foo';
+      const bar = 'bar';
+
+      const t = () => html`<div>${foo}${bar}</div>`;
+
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>foobar</div>'
+      );
+
+      foo = 'bbb';
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>bbbbar</div>'
+      );
+    });
+
+    test('renders and updates attributes', () => {
+      let foo = 'foo';
+      const bar = 'bar';
+
+      const t = () => html`<div a="${foo}:${bar}"></div>`;
+
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div a="foo:bar"></div>'
+      );
+
+      foo = 'bbb';
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div a="bbb:bar"></div>'
+      );
+    });
+
+    test('updates nested templates', () => {
+      let foo = 'foo';
+      const bar = 'bar';
+      const baz = 'baz';
+
+      const t = (x: boolean) => {
+        let partial;
+        if (x) {
+          partial = html`<h1>${foo}</h1>`;
+        } else {
+          partial = html`<h2>${bar}</h2>`;
+        }
+
+        return html`${partial}${baz}`;
+      };
+
+      render(t(true), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<h1>foo</h1>baz'
+      );
+
+      foo = 'bbb';
+      render(t(true), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<h1>bbb</h1>baz'
+      );
+
+      render(t(false), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<h2>bar</h2>baz'
+      );
+    });
+
+    test('updates an element', () => {
+      let child: any = document.createElement('p');
+      // prettier-ignore
+      const t = () => html`<div>${child}<div></div></div>`;
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div><p></p><div></div></div>'
+      );
+
+      child = undefined;
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div><div></div></div>'
+      );
+
+      child = document.createTextNode('foo');
+      render(t(), container);
+      assert.equal(
+        stripExpressionComments(container.innerHTML),
+        '<div>foo<div></div></div>'
+      );
+    });
+
+    test(
+      'overwrites an existing TemplateInstance if one exists and does ' +
+        'not have a matching Template',
+      () => {
+        render(html`<div>foo</div>`, container);
+
+        assert.equal(container.children.length, 1);
+        const fooDiv = container.children[0];
+        assert.equal(fooDiv.textContent, 'foo');
+
+        render(html`<div>bar</div>`, container);
+
+        assert.equal(container.children.length, 1);
+        const barDiv = container.children[0];
+        assert.equal(barDiv.textContent, 'bar');
+
+        assert.notEqual(fooDiv, barDiv);
+      }
+    );
   });
 });